refactor: introduce abstract base class for audit log handler
so we can share code for the audit log cleanup handler
